@@ -3,13 +3,11 @@ import type { ReactCSRPageConfig } from './page'
 export function getScriptString(page: ReactCSRPageConfig) {
   return `
 import React from 'react'
-import ReactDOM from 'react-dom'
-import { PageProvider } from 'xueyan-react'
+import { createRoot } from 'react-dom/client'
+import { PageProvider } from 'sdin-react'
 import Entry from ${JSON.stringify(page.entry)}
 
-ReactDOM.render(
-  <PageProvider Content={Entry}/>,
-  document.getElementById('app')
-)
+const root = createRoot(document.getElementById('app'))
+root.render(<PageProvider Content={Entry}/>)
   `
 }
