@@ -8,13 +8,13 @@ import type { ImageLinkProps, DocConfig } from 'xueyan-react-doc'
 import type { SelectOption } from 'xueyan-react-select'
 
 const ICON: ImageLinkProps = {
-  src: XT_ASSETS_PATH + 'project.png',
-  href: XT_PUBLIC_PATH
+  src: P_ASSETS_PATH + 'project.png',
+  href: P_PUBLIC_PATH
 }
 
 const ICONS: ImageLinkProps[] = [
   {
-    src: XT_ASSETS_PATH + 'repository.png',
+    src: P_ASSETS_PATH + 'repository.png',
     href: pkg.repository.url,
     title: 'repository',
   }
@@ -39,8 +39,8 @@ const LANGUAGES: SelectOption<string>[] = [
 ]
 
 const SIDE_FOOTER = `
-Author [${XT_AUTHOR_NAME}](mailto://${XT_AUTHOR_EMAIL})  
-Builder [xueyan-typescript-cli](https://github.com/xueyan-site/xueyan-typescript-cli)  
+Author [${P_AUTHOR_NAME}](mailto://${P_AUTHOR_EMAIL})  
+Builder [sdin](https://github.com/xueyan-site/sdin)  
 `
 
 export interface PageDocProps<T,D> extends 
