@@ -42,14 +42,14 @@ module.exports = {
         key: 'favicon',
         rel: "icon",
         type: "image/png",
-        href: "{{XT_ASSETS_PATH}}favicon.png"
+        href: "{{P_ASSETS_PATH}}favicon.png"
       }
     ],
     styles: [
       {
         key: 'global',
         rel: 'stylesheet',
-        href: '{{XT_ASSETS_PATH}}global.css'
+        href: '{{P_ASSETS_PATH}}global.css'
       }
     ],
     scripts: [
