@@ -50,7 +50,7 @@ function nodesToHTML(label: string, nodes: NodeAttrs[], variables: Record<string
 export function getTemplateString(
   root: string,
   pkg: PackageInfo,
-  pjtCfg: ReactCSRProjectConfig,
+  prjCfg: ReactCSRProjectConfig,
   pgCfg: ReactCSRPageConfig,
   dev: boolean
 ) {
@@ -60,24 +60,24 @@ export function getTemplateString(
     // 在开发模式下，需尽可能使用本地的包（减少CDN包，去除网络带来的阻碍）
     scripts = pgCfg.scripts.filter(i => !getDepVersion(pkg, i.key))
   }
-  const XT_PAGE = {
+  const P_PAGE = {
     id: pgCfg.id,
     name: pgCfg.name,
     path: pgCfg.path,
-    publicPath: pjtCfg.publicPath,
-    assetsPath: pjtCfg.assetsPath,
+    publicPath: prjCfg.publicPath,
+    assetsPath: prjCfg.assetsPath,
     privatePath: pgCfg.privatePath
   }
   const VARIABLES = {
-    XT_ID: pjtCfg.id, // 项目ID，一般是package.name
-    XT_TYPE: pjtCfg.type, // 项目类型，此处是react-csr
-    XT_NAME: pjtCfg.name, // 项目名称
-    XT_VERSION: pkg.version, // 项目版本
-    XT_AUTHOR: pkg.author, // 项目作者 author <email>
-    XT_AUTHOR_NAME: pkg.authorName, // 项目作者名称
-    XT_AUTHOR_EMAIL: pkg.authorEmail, // 项目作者邮箱
-    XT_PUBLIC_PATH: pjtCfg.publicPath, // 项目url中的公共路径（以'/'开头和结尾）
-    XT_ASSETS_PATH: pjtCfg.assetsPath, // 项目的素材路径（以'/'开头和结尾）
+    P_ID: prjCfg.id, // 项目ID，一般是package.name
+    P_TYPE: prjCfg.type, // 项目类型，此处是react-csr
+    P_NAME: prjCfg.name, // 项目名称
+    P_VERSION: pkg.version, // 项目版本
+    P_AUTHOR: pkg.author, // 项目作者 author <email>
+    P_AUTHOR_NAME: pkg.authorName, // 项目作者名称
+    P_AUTHOR_EMAIL: pkg.authorEmail, // 项目作者邮箱
+    P_PUBLIC_PATH: prjCfg.publicPath, // 项目url中的公共路径（以'/'开头和结尾）
+    P_ASSETS_PATH: prjCfg.assetsPath, // 项目的素材路径（以'/'开头和结尾）
   }
   return `
     <!DOCTYPE html>
@@ -86,14 +86,14 @@ export function getTemplateString(
         <meta charset="UTF-8"/>
         ${nodesToHTML('meta', pgCfg.metas, VARIABLES)}
         <title>${title}</title>
-        <script>window.XT_PAGE=${JSON.stringify(XT_PAGE)}</script>
+        <script>window.P_PAGE=${JSON.stringify(P_PAGE)}</script>
         ${nodesToHTML('link', pgCfg.links, VARIABLES)}
         ${nodesToHTML('script', scripts, VARIABLES)}
         ${nodesToHTML('link', pgCfg.styles, VARIABLES)}
       </head>
       <body>
         <div id="app">
-          ${pgCfg.skeleton(root, pkg, pjtCfg, pgCfg, dev)}
+          ${pgCfg.skeleton(root, pkg, prjCfg, pgCfg, dev)}
         </div>
       </body>
     </html>
